Move config types
