+++ src/config.ts
@@ -1,29 +1,4 @@
-import { z } from "zod";
-
-export type Config = z.infer<typeof configSchema>;
-
-export const method = z.enum(["GET", "POST", "PUT", "DELETE"]);
-
-export type Method = z.infer<typeof method>;
-
-const endpointSchema = z.object({
-  path: z.string(),
-  methods: z.array(method),
-  query: z.string(),
-});
-
-export type Endpoint = z.infer<typeof endpointSchema>;
-
-export const configSchema = z.object({
-  graphqlServer: z.object({
-    headers: z.object({
-      additional: z.record(z.string(), z.string()),
-      forward: z.array(z.string()),
-    }),
-  }),
-  headers: z.record(z.string(), z.string()),
-  restifiedEndpoints: z.array(endpointSchema),
-});
+import { Config } from "./modules/utils/types";
 
 // This configuration is used to configure the plugin when it is used
 // in development with Wrangler
